PR code clean up
