+++ common/src/main/java/net/william278/huskhomes/redis/RedisImpl.java
@@ -2,10 +2,10 @@ package net.william278.huskhomes.redis;
 
 
 import io.lettuce.core.RedisClient;
-import net.william278.huskhomes.redis.redisdata.RedisAbstract;
+import net.william278.huskhomes.redis.lettuce.RedisBase;
 
 
-public class RedisImpl extends RedisAbstract {
+public class RedisImpl extends RedisBase {
 
     public RedisImpl(RedisClient lettuceRedisClient) {
         super(lettuceRedisClient);
