+++ common/src/main/java/net/william278/huskhomes/redis/lettuce/RedisPubSub.java
@@ -1,0 +1,34 @@
+package net.william278.huskhomes.redis.lettuce;
+
+import io.lettuce.core.pubsub.RedisPubSubListener;
+
+public abstract class RedisPubSub<K,V> implements RedisPubSubListener<K,V> {
+    public abstract void message(K channel, V message);
+
+
+    @Override
+    public void message(K pattern, K channel, V message){
+
+    }
+
+
+    @Override
+    public void subscribed(K channel, long count){
+
+    }
+
+    @Override
+    public void psubscribed(K pattern, long count){
+
+    }
+
+    @Override
+    public void unsubscribed(K channel, long count){
+
+    }
+
+    @Override
+    public void punsubscribed(K pattern, long count){
+
+    }
+}
